code refactored and added methods to delete movie, actor, review

+++ moovi/src/main/java/com/dbms/project/moovi/business/service/AdminService.java
@@ -52,69 +52,112 @@ public class AdminService {
     @DeleteMapping("/api/delete/fan/{username}")
     public void deleteFan(
             @PathVariable("username") String username){
-        fanRepository.deleteById(fanRepository.findFanIdByUsername(username));
+        if(fanRepository.findById(fanRepository.findFanIdByUsername(username)).isPresent())
+            fanRepository.deleteById(fanRepository.findFanIdByUsername(username));
     }
 
     @DeleteMapping("/api/delete/critic/{username}")
     public void deleteCritic(
             @PathVariable("username") String username){
-        criticRepository.deleteById(criticRepository.findCriticIdByUsername(username));
+        if(criticRepository.findById(criticRepository.findCriticIdByUsername(username)).isPresent())
+            criticRepository.deleteById(criticRepository.findCriticIdByUsername(username));
     }
 
     @DeleteMapping("/api/delete/adrecruiter/{username}")
     public void deleteAdRecruiter(
             @PathVariable("username") String username){
-        adRecruiterRepository.deleteById(adRecruiterRepository.findAdRecruiterIdByUsername(username));
+        if(adRecruiterRepository.findById(adRecruiterRepository.findAdRecruiterIdByUsername(username)).isPresent())
+            adRecruiterRepository.deleteById(adRecruiterRepository.findAdRecruiterIdByUsername(username));
     }
 
     @DeleteMapping("/api/delete/theatremanager/{username}")
     public void deleteTheatreManager(
             @PathVariable("username") String username){
-        theatreManagerRepository.deleteById(theatreManagerRepository.findManagerIdByUsername(username));
+        if(theatreManagerRepository.findById(theatreManagerRepository.findManagerIdByUsername(username)).isPresent())
+            theatreManagerRepository.deleteById(theatreManagerRepository.findManagerIdByUsername(username));
     }
-    
+
+    @DeleteMapping("/api/delete/movie/{movieId}")
+    public void deleteMovie(
+            @PathVariable("movieId") long movieId){
+        if(movieRepository.findById(movieId).isPresent()){
+            movieRepository.deleteById(movieId);
+        }
+    }
+
+    @DeleteMapping("/api/delete/actor/{actorId}")
+    public void deleteActor(
+            @PathVariable("actorId") long actorId){
+        if(actorRepository.findById(actorId).isPresent()){
+            actorRepository.deleteById(actorId);
+        }
+    }
+
+    @DeleteMapping("/api/delete/review/{reviewId}")
+    public void deleteReview(
+            @PathVariable("reviewId") long reviewId){
+        if(reviewRepository.findById(reviewId).isPresent()){
+            reviewRepository.deleteById(reviewId);
+        }
+    }
+
     @PutMapping("/api/edit/fan/{username}")
     public Fan updateFan(
             @PathVariable("username") String username,
             @RequestBody Fan newFan) {
-    	Fan fan = fanRepository.findById(fanRepository.findFanIdByUsername(username)).get();
-    	fan.set(newFan);
-        return fanRepository.save(fan);
+        if(fanRepository.findById(fanRepository.findFanIdByUsername(username)).isPresent()) {
+            Fan fan = fanRepository.findById(fanRepository.findFanIdByUsername(username)).get();
+            fan.set(newFan);
+            return fanRepository.save(fan);
+        }
+        return null;
     }
-    
+
     @PutMapping("/api/edit/critic/{username}")
     public Critic updateCritic(
             @PathVariable("username") String username,
             @RequestBody Critic newCritic) {
-    	Critic critic = (Critic) criticRepository.findCriticByUsername(username);
-    	critic.set(newCritic);
-        return criticRepository.save(critic);
+        if(criticRepository.findById(criticRepository.findCriticIdByUsername(username)).isPresent()) {
+            Critic critic = criticRepository.findById(criticRepository.findCriticIdByUsername(username)).get();
+            critic.set(newCritic);
+            return criticRepository.save(critic);
+        }
+        return null;
     }
-    
+
     @PutMapping("/api/edit/adrecruiter/{username}")
     public AdRecruiter updateAdrecruiter(
             @PathVariable("username") String username,
             @RequestBody AdRecruiter newAdrecruiter) {
-    	AdRecruiter adr = (AdRecruiter) adRecruiterRepository.findAdRecruiterByUsername(username);
-    	adr.set(newAdrecruiter);
-        return adRecruiterRepository.save(adr);
+        if(adRecruiterRepository.findById(adRecruiterRepository.findAdRecruiterIdByUsername(username)).isPresent()) {
+            AdRecruiter adr = adRecruiterRepository.findById(adRecruiterRepository.findAdRecruiterIdByUsername(username)).get();
+            adr.set(newAdrecruiter);
+            return adRecruiterRepository.save(adr);
+        }
+        return null;
     }
-    
+
     @PutMapping("/api/edit/theatremanager/{username}")
     public TheatreManager updateTheatreManager(
             @PathVariable("username") String username,
             @RequestBody TheatreManager newtheatreManager) {
-    	TheatreManager theatreManager = (TheatreManager) theatreManagerRepository.findManagerByUsername(username);
-    	theatreManager.set(newtheatreManager);
-        return theatreManagerRepository.save(theatreManager);
+        if(theatreManagerRepository.findById(theatreManagerRepository.findManagerIdByUsername(username)).isPresent()) {
+            TheatreManager theatreManager = theatreManagerRepository.findById(theatreManagerRepository.findManagerIdByUsername(username)).get();
+            theatreManager.set(newtheatreManager);
+            return theatreManagerRepository.save(theatreManager);
+        }
+        return null;
     }
-    
+
     @PutMapping("/api/edit/theatre/{theatreId}")
     public Theatre updateTheatre(
             @PathVariable("theatreId") long theatreId,
             @RequestBody Theatre newTheatre) {
-    	Theatre theatre = (Theatre) theatreRepository.findTheatreById(theatreId);
-    	theatre.set(newTheatre);
-        return theatreRepository.save(theatre);
+        if(theatreRepository.findById(theatreId).isPresent()) {
+            Theatre theatre = theatreRepository.findById(theatreId).get();
+            theatre.set(newTheatre);
+            return theatreRepository.save(theatre);
+        }
+        return null;
     }
 }
